Checking before deleting an API - APISTORE-351
git-svn-id: https://svn.wso2.org/repos/wso2/carbon/platform/trunk@130771 a5903396-d722-0410-b921-86c7d4935375

+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/APIProviderImpl.java
@@ -452,7 +452,9 @@ class APIProviderImpl extends AbstractAPIManager implements APIProvider {
     private void removeFromGateway(API api) throws APIManagementException {
         try {
             RESTAPIAdminClient client = new RESTAPIAdminClient(api.getId());
-            client.deleteApi();
+            if (client.getApi() != null) {
+                client.deleteApi();
+            }
         } catch (AxisFault axisFault) {
             handleException("Error while creating new API in gateway", axisFault);
         }
